misc: Move Instance.backup to InstanceBackup
